Implement repository for movie-related data operations

+++ core/database/src/commonMain/kotlin/com/tmdb/database/entity/TrendingMovieEntity.kt
@@ -16,6 +16,6 @@ data class TrendingMovieEntity(
     val releaseDate: String?,
     val posterPath: String?,
     val backdropPath: String?,
-    val voteAverage: Double,
+    val voteAverage: Float,
     val fetchedAt: Long // timestamp in millis
 )
